Completes footer section for desktop size
Changes image in footer for better dimensions. Changes text container
in footer to have max size to ensure image does not overflow out of
container. Uses float on tablet layout to achieve text wrapping on about
picture. Adds z-index 1 to about picture to bring it in front of the
background color of the text card.